requirements.txt
absl-py==2.2.0
alembic==1.15.1
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
bcrypt==3.2.0
beautifulsoup4==4.13.3
billiard==4.2.1
bleach==6.2.0
blinker==1.9.0
Brlapi==0.8.3
cachetools==5.5.2
#catboost==1.2.7
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.1
chex==0.1.89
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
contourpy==1.3.1
cryptography==44.0.1
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
debugpy==1.8.13
decorator==5.2.1
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.7.0
duplicity==0.8.21
email_validator==2.2.0
etils==1.12.2
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.11
fastapi-cli==0.0.7
fasteners==0.14.1
fastjsonschema==2.21.1
filelock==3.18.0
flatbuffers==25.2.10
flax==0.10.4
fonttools==4.56.0
fqdn==1.5.1
fsspec==2025.3.0
future==0.18.3
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httplib2==0.20.2
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.3
humanize==4.12.1
idna==3.10
imageio==2.37.0
importlib-metadata==4.6.4
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.34.0
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.5.3
jaxlib==0.5.3
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kaggle==1.7.4.2
keras==3.9.0
keyring==23.5.0
kiwisolver==1.4.8
kombu==5.5.1
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
libclang==18.1.1
lightgbm==4.6.0
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
ml_dtypes==0.5.1
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.1.0
namex==0.0.8
narwhals==1.31.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.4.2
notebook==7.3.3
notebook_shim==0.2.4
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
olefile==0.46
opencv-python==4.11.0.86
opendatasets==0.1.22
opt_einsum==3.4.0
optax==0.2.4
optree==0.14.1
orbax-checkpoint==0.11.10
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
paramiko==3.3.2
parso==0.8.4
passlib==1.7.4
patsy==1.0.1
pexpect==4.8.0
pillow==11.1.0
platformdirs==4.3.7
plotly==6.0.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==2.10.6
pydantic-extra-types==2.10.3
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pytesseract==0.3.13
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-json-logger==3.3.0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.1
pyxdg==0.27
PyYAML==6.0.2
pyzmq==26.3.0
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
reportlab==3.6.13
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.23.1
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
SecretStorage==3.3.1
Send2Trash==1.8.3
sentencepiece==0.2.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.39
sqlmodel==0.0.24
stack-data==0.6.3
starlette==0.46.1
statsmodels==0.14.4
streamlit==1.43.2
sympy==1.13.1
systemd-python==234
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
tensorstore==0.1.72
termcolor==2.5.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.6.0
tifffile==2025.3.13
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
toolz==1.0.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
treescope==0.1.9
triton==3.2.0
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
ujson==5.10.0
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==2.3.0
usb-creator==0.3.7
uvicorn==0.34.0
uvloop==0.21.0
vine==5.1.0
wadllib==1.3.6
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
xdg==5
xgboost==3.0.0
xkit==0.0.0
zipp==3.21.0
